Move everything Command/exec.Cmd related to run package
