+++ dev/sg/internal/run/logger.go
@@ -1,4 +1,4 @@
-package main
+package run
 
 import (
 	"bytes"
