+++ dev/sg/internal/squash/squash.go
@@ -11,9 +11,9 @@ import (
 	"time"
 
 	"github.com/hashicorp/go-multierror"
-	"github.com/sourcegraph/sourcegraph/dev/sg/internal/command"
 	"github.com/sourcegraph/sourcegraph/dev/sg/internal/db"
 	"github.com/sourcegraph/sourcegraph/dev/sg/internal/migration"
+	"github.com/sourcegraph/sourcegraph/dev/sg/internal/run"
 	"github.com/sourcegraph/sourcegraph/dev/sg/internal/stdout"
 	"github.com/sourcegraph/sourcegraph/lib/output"
 )
@@ -90,8 +90,7 @@ func generateSquashedUpMigration(database db.Database, postgresDSN string) (_ st
 
 	pgDump := func(args ...string) (string, error) {
 		cmd := exec.Command("pg_dump", append([]string{postgresDSN}, args...)...)
-		cmd.Env = []string{}
-		return command.RunInRoot(cmd)
+		return run.InRoot(cmd)
 	}
 
 	pgDumpOutput, err := pgDump("--schema-only", "--no-owner", "--exclude-table", "*schema_migrations")
@@ -170,7 +169,7 @@ func runPostgresContainer(databaseName string) (_ func(err error) error, err err
 			"kill",
 			squasherContainerName,
 		}
-		if _, killErr := command.RunDocker(killArgs...); killErr != nil {
+		if _, killErr := run.DockerCmd(killArgs...); killErr != nil {
 			err = multierror.Append(err, fmt.Errorf("failed to stop docker container: %s", killErr))
 		}
 
@@ -185,7 +184,7 @@ func runPostgresContainer(databaseName string) (_ func(err error) error, err err
 		"-e", "POSTGRES_HOST_AUTH_METHOD=trust",
 		"postgres:12.6",
 	}
-	if _, err := command.RunDocker(runArgs...); err != nil {
+	if _, err := run.DockerCmd(runArgs...); err != nil {
 		return nil, err
 	}
 
@@ -200,7 +199,7 @@ func runPostgresContainer(databaseName string) (_ func(err error) error, err err
 		squasherContainerName,
 		"createdb", databaseName,
 	}
-	if _, err := command.RunDocker(execArgs...); err != nil {
+	if _, err := run.DockerCmd(execArgs...); err != nil {
 		return nil, teardown(err)
 	}
 
@@ -214,7 +213,7 @@ func runMigrate(database db.Database, postgresDSN string, args ...string) (strin
 		return "", err
 	}
 
-	return command.RunInRoot(exec.Command("migrate", append([]string{"-database", postgresDSN, "-path", baseDir}, args...)...))
+	return run.InRoot(exec.Command("migrate", append([]string{"-database", postgresDSN, "-path", baseDir}, args...)...))
 }
 
 // lastMigrationIndexAtCommit returns the index of the last migration for the given database
@@ -223,7 +222,7 @@ func runMigrate(database db.Database, postgresDSN string, args ...string) (strin
 func lastMigrationIndexAtCommit(database db.Database, commit string) (int, bool, error) {
 	migrationsDir := filepath.Join("migrations", database.Name)
 
-	output, err := command.RunGit("ls-tree", "-r", "--name-only", commit, migrationsDir)
+	output, err := run.GitCmd("ls-tree", "-r", "--name-only", commit, migrationsDir)
 	if err != nil {
 		return 0, false, err
 	}
